[fix] update validators for new device type
